DRF_Example/urls.py

from django.contrib import admin

from django.conf import settings

from django.urls import include, path
from rest_framework import routers
from DRF_App import views

router = routers.DefaultRouter()
router.register(r'users', views.UserViewSet)
router.register(r'groups', views.GroupViewSet)
router.register(r'City', views.CityViewSet)


urlpatterns = [
    path('admin/', admin.site.urls),
    path('', include(router.urls)),
    path('api-auth/', include('rest_framework.urls', namespace='rest_framework'))

]