requirements.txt
amqp==5.0.2
appdirs==1.4.4
apturl==0.5.2
asgiref==3.2.10
astroid==2.1.0
asttokens==2.0.3
attrs==20.3.0
autopep8==1.4.4
bcrypt==3.1.7
beautifulsoup4==4.8.2
billiard==3.6.3.0
bleach==3.2.1
blinker==1.4
Brlapi==0.7.0
cached-property==1.5.2
cairocffi==1.2.0
CairoSVG==2.5.0
celery==5.0.2
certifi==2019.9.11
cffi==1.14.4
chardet==3.0.4
chrome-gnome-shell==0.0.0
Click==7.0
click-didyoumean==0.0.3
click-repl==0.1.6
colorama==0.4.3
command-not-found==0.3
cryptography==2.8
cssselect2==0.4.1
cupshelpers==1.0
cycler==0.10.0
dateparser==0.7.2
dbus-python==1.2.16
decouple==0.0.7
defer==1.0.6
defusedxml==0.6.0
distlib==0.3.1
distro==1.4.0
distro-info===0.23ubuntu1
Django==3.0.7
django-admin==2.0.1
django-cleanup==5.1.0
django-crispy-forms==1.8.0
django-debug-permissions==1.0.0
django-debug-toolbar==2.1
django-discover-runner==1.0
django-environ==0.4.5
django-excel-base==1.0.4
django-excel-response2==3.0.2
django-extensions==3.0.2
django-makemessages-xgettext==0.1.1
django-model-utils==4.0.0
django-role-permissions==2.2.1
django-six==1.0.4
django-taggit==1.3.0
django-tinymce==3.1.0
django-utils-six==2.0
django-widget-tweaks==1.4.8
djangorestframework==3.12.2
docutils==0.16
duplicity==0.8.12.0
entrypoints==0.3
fake-factory==9999.9.9
Faker==2.0.4
fasteners==0.14.1
filelock==3.0.12
flake8==3.8.4
future==0.18.2
graphviz==0.14
html5lib==1.0.1
httplib2==0.14.0
idna==2.8
install==1.3.4
isodate==0.6.0
isort==4.3.4
jedi==0.15.2
jet-bridge-base==0.7.7
jet-django==1.0.2
kazam==1.4.5
keyring==18.0.1
kiwisolver==1.3.1
kombu==5.0.2
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lazy-object-proxy==1.3.1
lockfile==0.12.2
logilab-common==1.4.3
louis==3.12.0
lxml==4.5.0
macaroonbakery==1.3.1
Mako==1.1.0
MarkupSafe==1.1.0
matplotlib==3.3.3
mccabe==0.6.1
monotonic==1.5
mypy==0.761
mypy-extensions==0.4.3
mysql-connector-python==8.0.20
mysqlclient==2.0.1
netifaces==0.10.4
netsuitesdk==1.12.0
numpy==1.19.4
oauthlib==3.1.0
olefile==0.46
packaging==20.7
paramiko==2.6.0
parso==0.5.2
pbr==5.5.1
pexpect==4.6.0
Pillow==7.2.0
prompt-toolkit==2.0.9
protobuf==3.12.2
psutil==5.5.1
pycairo==1.16.2
pycodestyle==2.5.0
pycparser==2.19
pycups==1.9.73
pydot==1.4.1
pyflakes==2.2.0
Pygments==2.3.1
PyGObject==3.36.0
PyJWT==1.7.1
pylint==2.4.4
pymacaroons==0.13.0
PyNaCl==1.3.0
pyodbc==4.0.30
pyparsing==2.4.7
Pyphen==0.10.0
pyRFC3339==1.1
pyserial==3.4
python-apt==2.0.0+ubuntu0.20.4.1
python-dateutil==2.8.1
python-debian===0.1.36ubuntu1
python-decouple==3.3
pytz==2020.1
pyxdg==0.26
PyYAML==5.3.1
regex==2019.11.1
reportlab==3.5.34
requests==2.22.0
requests-file==1.5.1
requests-toolbelt==0.9.1
requests-unixsocket==0.2.0
roman==2.0.0
screen==1.0.1
SecretStorage==2.3.1
Send2Trash==1.5.0
simple-http-server==0.3.1
simplejson==3.16.0
six==1.14.0
soupsieve==1.9.5
SQLAlchemy==1.3.11
sqlparse==0.3.1
suds-jurko==0.6
systemd-python==234
testresources==2.0.1
text-unidecode==1.3
thonny==3.2.7
tinycss2==1.1.0
typed-ast==1.4.1
typing-extensions==3.7.4.1
tzlocal==2.0.0
ubuntu-advantage-tools==20.3
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.25.7
usb-creator==0.3.7
vine==5.0.0
virtualenv==20.1.0
wadllib==1.3.3
wcwidth==0.1.7
WeasyPrint==52.1
webencodings==0.5.1
wrapt==1.11.1
xkit==0.0.0
XlsxWriter==1.3.7
xlwt==1.3.0
zeep==4.0.0
